refactor(link-embedder): separate imports
