+++ src/components/cards/LinkEmbedder/LinkEmbedder.skeleton.tsx
@@ -1,4 +1,5 @@
-import { Card, CardBody, Skeleton } from "@nextui-org/react";
+import { Card, CardBody } from "@nextui-org/card";
+import { Skeleton } from "@nextui-org/skeleton";
 
 import { cn } from "@/utils/cn";
